move add set button up to Exercise screen rather than ExerciseLog
